Fixed id returned as string + added controller & route for user password reset.

+++ src/balances/model/balance.model.ts
@@ -18,7 +18,7 @@ const { db_schema } = config;
 export default class Balance extends S.Model implements BalanceModel {
   @S.PrimaryKey
 	@S.AutoIncrement
-	@S.Column(S.DataType.BIGINT)
+	@S.Column(S.DataType.INTEGER)
 	public id!: number;
 
 	@S.AllowNull(false)
