+++ src/feedbacks/model/feedback.model.ts
@@ -20,7 +20,7 @@ const { db_schema } = config;
 export default class Feedback extends S.Model implements FeedbackModel {
   @S.PrimaryKey
 	@S.AutoIncrement
-	@S.Column(S.DataType.BIGINT)
+	@S.Column(S.DataType.INTEGER)
 	public id!: number;
 
 	@S.AllowNull(false)
